Separating out the polyfill code so as to make the file(s) readable and reduce the size
